fix bugs

+++ tests/test_checkpoint_hook.py
@@ -17,7 +17,6 @@ from hurricane.hooks import CheckpointHook
 temp_folder_path = Path(__file__).parents[1] / '_temp_checkpoints'
 
 
-
 class _TestTrainer(Trainer):
     def __init__(self):
         model = nn.Linear(1, 1)
@@ -42,21 +41,24 @@ class _TestTrainer(Trainer):
 
 def test_checkpoint_hook_dataloader():
     # set up test folder
-    if temp_folder_path.exists():
-        shutil.rmtree(temp_folder_path)
-    temp_folder_path.mkdir(parents=True)
-    
+    temp_folder_path.mkdir(parents=True, exist_ok=True)
     trainer = _TestTrainer()
+    if trainer.accelerator.is_main_process:
+        shutil.rmtree(temp_folder_path)
+        temp_folder_path.mkdir(parents=True)
+    trainer.accelerator.wait_for_everyone()
+    
     trainer.run()
     assert len(trainer.iterated_results) == 20, "Not all batches are iterated."
     original_results = trainer.iterated_results.copy()
-    del trainer
     
     # remove all but the 5th and the 15th
-    ckpt_dirs = [d for d in temp_folder_path.iterdir() if d.is_dir() and d.name.startswith('ckpt_step_')]
-    for ckpt_dir in ckpt_dirs:
-        if ckpt_dir.name not in ['ckpt_step_5', 'ckpt_step_15']:
-            shutil.rmtree(ckpt_dir)
+    if trainer.accelerator.is_main_process:
+        ckpt_dirs = [d for d in temp_folder_path.iterdir() if d.is_dir() and d.name.startswith('ckpt_step_')]
+        for ckpt_dir in ckpt_dirs:
+            if ckpt_dir.name not in ['ckpt_step_5', 'ckpt_step_15']:
+                shutil.rmtree(ckpt_dir)
+    trainer.accelerator.wait_for_everyone()
     
     # test reproducibility on the 2nd epoch
     trainer = _TestTrainer()
@@ -64,13 +66,14 @@ def test_checkpoint_hook_dataloader():
     assert len(trainer.iterated_results) == 5, "Continued number of batches is not correct."
     new_results = trainer.iterated_results.copy()
     assert new_results == original_results[15:], "Continued batches do not match the original."
-    del trainer
 
     # remove all but the 5th
-    ckpt_dirs = [d for d in temp_folder_path.iterdir() if d.is_dir() and d.name.startswith('ckpt_step_')]
-    for ckpt_dir in ckpt_dirs:
-        if ckpt_dir.name not in ['ckpt_step_5']:
-            shutil.rmtree(ckpt_dir)
+    if trainer.accelerator.is_main_process:
+        ckpt_dirs = [d for d in temp_folder_path.iterdir() if d.is_dir() and d.name.startswith('ckpt_step_')]
+        for ckpt_dir in ckpt_dirs:
+            if ckpt_dir.name not in ['ckpt_step_5']:
+                shutil.rmtree(ckpt_dir)
+    trainer.accelerator.wait_for_everyone()
     
     # test reproducibility on the 1st epoch
     trainer = _TestTrainer()
@@ -78,67 +81,68 @@ def test_checkpoint_hook_dataloader():
     assert len(trainer.iterated_results) == 15, "Continued number of batches is not correct."
     new_results = trainer.iterated_results.copy()
     assert new_results == original_results[5:], "Continued batches do not match the original."
-    del trainer
     
     # clean up
-    shutil.rmtree(temp_folder_path)
-
+    if trainer.accelerator.is_main_process:
+        shutil.rmtree(temp_folder_path)
+    trainer.accelerator.wait_for_everyone()
+    
 
 def test_checkpoint_hook_context():
     # set up test folder
-    if temp_folder_path.exists():
-        shutil.rmtree(temp_folder_path)
-    temp_folder_path.mkdir(parents=True)
-    
+    temp_folder_path.mkdir(parents=True, exist_ok=True)
     trainer = _TestTrainer()
+    if trainer.accelerator.is_main_process:
+        shutil.rmtree(temp_folder_path)
+        temp_folder_path.mkdir(parents=True)
+    trainer.accelerator.wait_for_everyone()
+        
     trainer.run()
     original_context = deepcopy(trainer.ctx)
-    del trainer
     
     # remove all but the 5th and the 15th
-    ckpt_dirs = [d for d in temp_folder_path.iterdir() if d.is_dir() and d.name.startswith('ckpt_step_')]
-    for ckpt_dir in ckpt_dirs:
-        if ckpt_dir.name not in ['ckpt_step_5', 'ckpt_step_15']:
-            shutil.rmtree(ckpt_dir)
+    if trainer.accelerator.is_main_process:
+        ckpt_dirs = [d for d in temp_folder_path.iterdir() if d.is_dir() and d.name.startswith('ckpt_step_')]
+        for ckpt_dir in ckpt_dirs:
+            if ckpt_dir.name not in ['ckpt_step_5', 'ckpt_step_15']:
+                shutil.rmtree(ckpt_dir)
+    trainer.accelerator.wait_for_everyone()
     
     # test reproducibility on the 2nd epoch
     trainer = _TestTrainer()
     trainer.run()
     new_context = deepcopy(trainer.ctx)
     for key in vars(original_context).keys():
-        if key not in ['step_loss', 'batches']:
-            new_value = getattr(new_context, key)
-            original_value = getattr(original_context, key)
-            assert new_value == original_value, \
-                (
-                    f"Context key {key} does not match.\n"
-                    f"\tBefore: {original_value}\n"
-                    f"\tAfter: {new_value}\n"
-                )
-                
-    del trainer
+        new_value = getattr(new_context, key)
+        original_value = getattr(original_context, key)
+        assert new_value == original_value, \
+            (
+                f"Context key {key} does not match.\n"
+                f"\tBefore: {original_value}\n"
+                f"\tAfter: {new_value}\n"
+            )
 
     # remove all but the 5th
-    ckpt_dirs = [d for d in temp_folder_path.iterdir() if d.is_dir() and d.name.startswith('ckpt_step_')]
-    for ckpt_dir in ckpt_dirs:
-        if ckpt_dir.name not in ['ckpt_step_5']:
-            shutil.rmtree(ckpt_dir)
+    if trainer.accelerator.is_main_process:
+        ckpt_dirs = [d for d in temp_folder_path.iterdir() if d.is_dir() and d.name.startswith('ckpt_step_')]
+        for ckpt_dir in ckpt_dirs:
+            if ckpt_dir.name not in ['ckpt_step_5']:
+                shutil.rmtree(ckpt_dir)
+    trainer.accelerator.wait_for_everyone()
     
     # test reproducibility on the 1st epoch
     trainer = _TestTrainer()
     trainer.run()
     new_context = deepcopy(trainer.ctx)
     for key in vars(original_context).keys():
-        if key not in ['step_loss', 'batches']:
-            assert new_value == original_value, \
-                (
-                    f"Context key {key} does not match.\n"
-                    f"\tBefore: {original_value}\n"
-                    f"\tAfter: {new_value}\n"
-                )
-    del trainer
+        assert new_value == original_value, \
+            (
+                f"Context key {key} does not match.\n"
+                f"\tBefore: {original_value}\n"
+                f"\tAfter: {new_value}\n"
+            )
     
     # clean up
-    shutil.rmtree(temp_folder_path)
-
-test_checkpoint_hook_context()
+    if trainer.accelerator.is_main_process:
+        shutil.rmtree(temp_folder_path)
+    trainer.accelerator.wait_for_everyone()
